Rename module to curve from characteristic
