+++ sympy/physics/_biomechanics/curve.py
